Allow to add objects to generic text.
Up to now generic text accepted only text. This is a problem, because if I have
a multi-line text the resulting indentation is wrong: GenericCodeBlock attaches
an indentation before the full string , if there are internal indentations they
don't follow the main one.

Giving the possibility to add an object allows to use the code() method of the
specific object. In that way is the object itself to manage its own indentation.